added Hilt DI

+++ Tutor/app/src/main/java/com/composeplayground/tutor/domain/entities/CourseEntity.kt
@@ -1,6 +1,6 @@
 package com.composeplayground.tutor.domain.entities
 
-data class CourseEntity(val courseId: Int,
+data class CourseEntity(val courseId: Long,
                         val courseName: String,
                         val courseRating: Int,
                         val facilitator: Facilitator)
