+++ Tutor/app/src/main/java/com/composeplayground/tutor/domain/entities/Facilitator.kt
@@ -1,4 +1,4 @@
 package com.composeplayground.tutor.domain.entities
 
-data class Facilitator(val id: Int,
+data class Facilitator(val id: Long,
                        val name: String)
